Revert CHANGELOG edits from commit 6995982a
Undoing the spelling corrections to the CHANGELOG as it accurately
reflects the name of the file at the time the CHANGELOG was written. If
someone were to view an older version of the code, they should expect the
filename to be misspelled. We will make a note of the spelling
correction in the next release.