Rename ./bin to ./rsm
This changed was motivated by a difficult to diagnose problem wherein
GSMI was referencing the "bin" package of another python package, namely
"redi", which caused the code to fail.

It was decided to avoid a name conflict like this in the future, we
should use project-specific names for our source-files directory.

Note: I used PyCharm to rename the folder, double-checking the changes
it suggested; re-ran the unit tests; then, ran GSMI and GSM with our
development configuration.

Closes ctsit/research-subject-mapper#15.